littlefs: Fix Information.Offset multiplication overflow BlockSize and i are both uint32_t and should be upcasted first.
Issue: GHSA-7hqh-mq57-wjmq
Reported-by: Jaroslav Lobačevski <jarlob@github.com>

+++ NanaZip.Codecs/NanaZip.Codecs.Archive.Littlefs.cpp
@@ -648,7 +648,7 @@ namespace NanaZip::Codecs::Archive
                     Information.Type = LfsTypeRegular;
                     Information.Size = BlockSize;
                     Information.Path = Mile::FormatString("[%d]", i);
-                    Information.Offset = i * BlockSize;
+                    Information.Offset = static_cast<std::uint64_t>(i) * BlockSize;
                     this->m_FilePaths.emplace_back(Information);
                 }
